[SPARK-46507][PS][TESTS] Split `IndexesSlowTests` into multiple tests
### What changes were proposed in this pull request?
Split `IndexesSlowTests` into multiple tests

### Why are the changes needed?
for testing parallelism

### Does this PR introduce _any_ user-facing change?
no, test-only

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44491 from zhengruifeng/ps_test_break_idx_base_slow.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
