+++ dev/sparktestsupport/modules.py
@@ -801,7 +801,10 @@ pyspark_pandas_slow = Module(
         "pyspark.pandas.tests.indexes.test_diff",
         "pyspark.pandas.tests.indexes.test_insert",
         "pyspark.pandas.tests.indexes.test_map",
-        "pyspark.pandas.tests.indexes.test_base_slow",
+        "pyspark.pandas.tests.indexes.test_append",
+        "pyspark.pandas.tests.indexes.test_intersection",
+        "pyspark.pandas.tests.indexes.test_monotonic",
+        "pyspark.pandas.tests.indexes.test_union",
         "pyspark.pandas.tests.indexes.test_datetime",
         "pyspark.pandas.tests.indexes.test_datetime_at",
         "pyspark.pandas.tests.indexes.test_datetime_between",
@@ -1183,7 +1186,10 @@ pyspark_pandas_connect_part2 = Module(
         "pyspark.pandas.tests.connect.computation.test_parity_pivot_table_multi_idx",
         "pyspark.pandas.tests.connect.computation.test_parity_pivot_table_multi_idx_adv",
         "pyspark.pandas.tests.connect.computation.test_parity_stats",
-        "pyspark.pandas.tests.connect.indexes.test_parity_base_slow",
+        "pyspark.pandas.tests.connect.indexes.test_parity_append",
+        "pyspark.pandas.tests.connect.indexes.test_parity_intersection",
+        "pyspark.pandas.tests.connect.indexes.test_parity_monotonic",
+        "pyspark.pandas.tests.connect.indexes.test_parity_union",
         "pyspark.pandas.tests.connect.frame.test_parity_interpolate",
         "pyspark.pandas.tests.connect.frame.test_parity_interpolate_error",
         "pyspark.pandas.tests.connect.series.test_parity_interpolate",
